// src/types.ts
export interface User {
  id: number;
  name: string;
  username?: string;
  email: string;
  address: Address;
  phone?: string;
  website?: string;
  company?: Company;
  isAllTasksCompleted?: boolean;
}
export interface Address {
  street: string;
  suite?: string;
  city: string;
  zipcode: string;
  geo?: Geo;
}
export interface Geo {
  lat: string;
  lng: string;
}
export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface Post {
  userId: number;
  id: number;
  title: string;
  body: string;
}

export interface ToDo {
  userId: number;
  id: number;
  title: string;
  completed: boolean;
}